pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mosaic-photo-generator"
version = "0.2.0"
description = "Photo mosaic generator: tiles a target image with closest-RGB-match thumbnails."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.0",
    "pillow>=10.0",
]

[project.scripts]
mosaic = "mosaic.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
]

[tool.hatch.build.targets.wheel]
packages = ["src/mosaic"]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B011"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
